feat: Implement item deletion functionality in CarritoCompras
The `CarritoCompras` component has been updated to include the functionality to delete items from the shopping cart. This is achieved by adding the `EliminarItem` function, which removes the selected item from the `sesionCarrito` state using the `deleteItem` action. After deleting the item, the user is redirected to the `/Carriot` page.