feat: Update CarritoActions.js and Detalles.js
This commit updates the `CarritoActions.js` file to use the `HttpCliente` object instead of `instancia` for making API requests. It also adds a console log statement for debugging purposes.

In addition, the `Detalles.js` file is modified to fix a typo in the `productoSeleccionado` object. The `_id` property is corrected to `id`.