feat: Refactor CarritoActions.js and Detalles.js
This commit refactors the `CarritoActions.js` file to use the `HttpCliente` object for making API requests instead of `instancia`. It also removes unnecessary console log statements for debugging purposes.

In addition, the `Detalles.js` file is modified to fix a typo in the `productoSeleccionado` object. The `_id` property is corrected to `id`.